Transform EntityDefinition to support both import and export
