+++ src/commands/export.ts
@@ -5,20 +5,20 @@ import JSZip from "jszip";
 import packageFile from "../../package.json" assert { type: "json" };
 import { RegisterCommand } from "../types/yargs.js";
 import { serially } from "../utils/requests.js";
-import { assetFoldersExportEntity } from "./export/entities/assetFolders.js";
-import { assetsExportEntity } from "./export/entities/assets.js";
-import { collectionsExportEntity } from "./export/entities/collections.js";
-import { contentItemsExportEntity } from "./export/entities/contentItems.js";
-import { contentTypesExportEntity } from "./export/entities/contentTypes.js";
-import { contentTypesSnippetsExportEntity } from "./export/entities/contentTypesSnippets.js";
-import { languagesExportEntity } from "./export/entities/languages.js";
-import { languageVariantsExportEntity } from "./export/entities/languageVariants.js";
-import { previewUrlsExportEntity } from "./export/entities/previewUrls.js";
-import { rolesExportEntity } from "./export/entities/roles.js";
-import { spacesExportEntity } from "./export/entities/spaces.js";
-import { taxonomiesExportEntity } from "./export/entities/taxonomies.js";
-import { workflowsExportEntity } from "./export/entities/workflows.js";
-import { EntityDefinition } from "./export/entityDefinition.js";
+import { assetFoldersExportEntity } from "./importExportEntities/entities/assetFolders.js";
+import { assetsExportEntity } from "./importExportEntities/entities/assets.js";
+import { collectionsExportEntity } from "./importExportEntities/entities/collections.js";
+import { contentItemsExportEntity } from "./importExportEntities/entities/contentItems.js";
+import { contentTypesExportEntity } from "./importExportEntities/entities/contentTypes.js";
+import { contentTypesSnippetsExportEntity } from "./importExportEntities/entities/contentTypesSnippets.js";
+import { languagesExportEntity } from "./importExportEntities/entities/languages.js";
+import { languageVariantsExportEntity } from "./importExportEntities/entities/languageVariants.js";
+import { previewUrlsExportEntity } from "./importExportEntities/entities/previewUrls.js";
+import { rolesExportEntity } from "./importExportEntities/entities/roles.js";
+import { spacesExportEntity } from "./importExportEntities/entities/spaces.js";
+import { taxonomiesExportEntity } from "./importExportEntities/entities/taxonomies.js";
+import { workflowsExportEntity } from "./importExportEntities/entities/workflows.js";
+import { EntityDefinition } from "./importExportEntities/entityDefinition.js";
 
 const zip = new JSZip();
