--- src/commands/export/entities/previewUrls.ts
@@ -1,9 +1,0 @@
-import { PreviewContracts } from "@kontent-ai/management-sdk";
-
-import { EntityDefinition } from "../entityDefinition.js";
-
-export const previewUrlsExportEntity: EntityDefinition<PreviewContracts.IPreviewConfigurationContract> = {
-    name: "previewUrls",
-    fetchEntities: client => client.getPreviewConfiguration().toPromise().then(res => res.rawData),
-    serializeEntities: collections => JSON.stringify(collections),
-  };
