+++ src/commands/importExportEntities/entities/collections.ts
@@ -6,4 +6,6 @@ export const collectionsExportEntity: EntityDefinition<ReadonlyArray<CollectionC
   name: "collections",
   fetchEntities: client => client.listCollections().toPromise().then(res => res.rawData.collections),
   serializeEntities: collections => JSON.stringify(collections),
+  importEntities: () => { throw new Error("Not supported yet.")},
+  deserializeEntities: () => { throw new Error("Not supported yet.")},
 };
