+++ src/commands/importExportEntities/entities/languageVariants.ts
@@ -19,4 +19,6 @@ export const languageVariantsExportEntity: EntityDefinition<ReadonlyArray<Langua
     return variants.flatMap(arr => arr);
   },
   serializeEntities: collections => JSON.stringify(collections),
+  importEntities: () => { throw new Error("Not supported yet.")},
+  deserializeEntities: () => { throw new Error("Not supported yet.")},
 };
