+++ src/commands/importExportEntities/entities/spaces.ts
@@ -6,4 +6,6 @@ export const spacesExportEntity: EntityDefinition<ReadonlyArray<SpaceContracts.I
   name: "spaces",
   fetchEntities: client => client.listSpaces().toPromise().then(res => res.rawData),
   serializeEntities: spaces => JSON.stringify(spaces),
+  importEntities: () => { throw new Error("Not supported yet.")},
+  deserializeEntities: () => { throw new Error("Not supported yet.")},
 }
