+++ src/commands/importExportEntities/entities/taxonomies.ts
@@ -6,4 +6,6 @@ export const taxonomiesExportEntity: EntityDefinition<ReadonlyArray<TaxonomyCont
   name: "taxonomies",
   fetchEntities: client => client.listTaxonomies().toAllPromise().then(res => res.data.items.map(t => t._raw)),
   serializeEntities: taxonomies => JSON.stringify(taxonomies),
+  importEntities: () => { throw new Error("Not supported yet.")},
+  deserializeEntities: () => { throw new Error("Not supported yet.")},
 };
