+++ src/commands/importExportEntities/entities/workflows.ts
@@ -1,0 +1,11 @@
+import { WorkflowContracts } from "@kontent-ai/management-sdk";
+
+import { EntityDefinition } from "../entityDefinition.js";
+
+export const workflowsExportEntity: EntityDefinition<WorkflowContracts.IListWorkflowsResponseContract> = {
+  name: "workflows",
+  fetchEntities: client => client.listWorkflows().toPromise().then(res => res.rawData),
+  serializeEntities: collections => JSON.stringify(collections),
+  importEntities: () => { throw new Error("Not supported yet.")},
+  deserializeEntities: () => { throw new Error("Not supported yet.")},
+};
